+++ pkgs/native_assets_builder/test/helpers.dart
@@ -104,7 +104,7 @@ Uri findPackageRoot(String packageName) {
 
 final pkgNativeAssetsBuilderUri = findPackageRoot('native_assets_builder');
 
-final testDataUri = pkgNativeAssetsBuilderUri.resolve('test/data/');
+final testDataUri = pkgNativeAssetsBuilderUri.resolve('test_data/');
 
 extension on Uri {
   String get name => pathSegments.where((e) => e != '').last;
@@ -153,7 +153,7 @@ Future<void> copyTestProjects({
     final targetFileUri = targetUri.resolveUri(pathToModify);
     final sourceString = await sourceFile.readAsString();
     final modifiedString = sourceString.replaceAll(
-      'path: ../../../',
+      'path: ../../',
       'path: ${pkgNativeAssetsBuilderUri.toFilePath().replaceAll('\\', '/')}',
     );
     await File.fromUri(targetFileUri)